Handles the default site through the intents extension.
